+++ src/components/form/date-time/date-time-scroller.component.tsx
@@ -1,11 +1,14 @@
 import { TypeOrArray } from '@beesoft/common';
-import { ReactNode, useContext } from 'react';
+import { Dispatch, ReactNode, useContext } from 'react';
 import { BeeSoftIcon } from '../../common/beesoft-icon/beesoft-icon.component.tsx';
 import { IconSize } from '../../common/beesoft-icon/beesoft-icon.props.ts';
-import TemplateOutlet, { TemplateFunction } from '../../common/template-outlet/template-outlet.component';
+import TemplateOutlet from '../../common/template-outlet/template-outlet.component';
 import { Button } from '../../navigation/buttons/button/button.component.tsx';
 import { DateTimeContext } from './date-time-context';
 import { DateScrollerType } from './date-time-types';
+import { useAddDateTimeBaseTemplateProps } from './hooks/add-date-time-base-template-props.hook.ts';
+import { DateTimeReducerAction } from './date-time.reducer.ts';
+import { DateTimeScrollerTemplateProps } from './date-time.props.ts';
 
 export interface DateTimeScrollerProps {
   title: string;
@@ -13,17 +16,9 @@ export interface DateTimeScrollerProps {
   onTitleClicked?: () => void;
   onMovePrevious: () => void;
   onMoveNext: () => void;
+  viewDate: Date;
+  dispatcher: Dispatch<DateTimeReducerAction>;
 }
-
-export interface DateTimeScrollerTemplateProps {
-  title: string;
-  scrollerType: DateScrollerType;
-  onTitleClicked?: () => void;
-  onMovePrevious: () => void;
-  onMoveNext: () => void;
-}
-
-export type DateTimeScrollerTemplate = TemplateFunction<DateTimeScrollerTemplateProps>;
 
 const DateTimeScroller = ({
   title,
@@ -31,17 +26,23 @@ const DateTimeScroller = ({
   onTitleClicked,
   onMovePrevious,
   onMoveNext,
+  viewDate,
+  dispatcher,
 }: DateTimeScrollerProps) => {
   const context = useContext(DateTimeContext);
   const viewTemplate = context.dateScrollerTemplate;
 
-  const templateProps: DateTimeScrollerTemplateProps = {
-    title,
-    scrollerType,
-    onTitleClicked,
-    onMovePrevious,
-    onMoveNext,
-  };
+  const templateProps = useAddDateTimeBaseTemplateProps<DateTimeScrollerTemplateProps>(
+    {
+      title,
+      scrollerType,
+      onTitleClicked,
+      onMovePrevious,
+      onMoveNext,
+    },
+    viewDate,
+    dispatcher
+  );
 
   const defaultTemplate = (_props: DateTimeScrollerTemplateProps, children: TypeOrArray<ReactNode>) => (
     <div className="bc-dt-scroller bsc-flex bsc-w-full bsc-flex-row bsc-items-center bsc-px-2 bsc-py-1">{children}</div>